backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.118.0
idna==3.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
